requirements.txt
# Core MongoDB
pymongo==4.6.1
dnspython==2.4.2

# Ollama Integration
ollama==0.1.7
requests==2.31.0

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
# Environment
python-dotenv==1.0.0

# Production
gunicorn==21.2.0

